rename rendering/svg/SVGPathData.{h,cpp} to SVGPathFromElement.{h,cpp} https://bugs.webkit.org/show_bug.cgi?id=316966 rdar://179478197
Reviewed by Nikolas Zimmermann.

This just renames SVGPathData to SVGPathFromElement to avoid
a name clash when it is time to implement Mixin SVGPathData.
https://w3c.github.io/svgwg/specs/paths/#InterfaceSVGPathData

* Source/WebCore/SaferCPPExpectations/UncheckedLocalVarsCheckerExpectations:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/rendering/PathOperation.cpp:
* Source/WebCore/rendering/svg/RenderSVGModelObject.cpp:
* Source/WebCore/rendering/svg/RenderSVGShape.cpp:
* Source/WebCore/rendering/svg/RenderSVGTextPath.cpp:
* Source/WebCore/rendering/svg/SVGPathFromElement.cpp: Renamed from Source/WebCore/rendering/svg/SVGPathData.cpp.
(WebCore::pathFromCircleElement):
(WebCore::pathFromEllipseElement):
(WebCore::pathFromLineElement):
(WebCore::pathFromPathElement):
(WebCore::pathFromPolygonElement):
(WebCore::pathFromPolylineElement):
(WebCore::pathFromRectElement):
(WebCore::pathFromUseElement):
(WebCore::pathFromGraphicsElement):
* Source/WebCore/rendering/svg/SVGPathFromElement.h: Renamed from Source/WebCore/rendering/svg/SVGPathData.h.
* Source/WebCore/rendering/svg/legacy/LegacyRenderSVGShape.cpp:
* Source/WebCore/svg/SVGAnimateMotionElement.cpp:
* Source/WebCore/svg/SVGGraphicsElement.cpp:

Canonical link: https://commits.webkit.org/315149@main
