+++ Source/WebCore/rendering/svg/SVGPathFromElement.h
