+++ Source/WebCore/svg/SVGAnimateMotionElement.cpp
@@ -34,8 +34,8 @@
 #include "SVGMPathElement.h"
 #include "SVGNames.h"
 #include "SVGParserUtilities.h"
-#include "SVGPathData.h"
 #include "SVGPathElement.h"
+#include "SVGPathFromElement.h"
 #include "SVGPathUtilities.h"
 #include "Settings.h"
 #include "StyleUpdate.h"
